requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
chromadb
pypdf
streamlit
python-dotenv
sentence-transformers